refactor: change weight field type to IntegerField in AthleteInfo model

+++ app_run/models.py
@@ -16,5 +16,5 @@ class Run(models.Model):
 class AthleteInfo(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, primary_key=True, related_name='athlete_info')
     goals = models.TextField(blank=True, default='')
-    weight = models.FloatField(blank=True, null=True)
+    weight = models.IntegerField(blank=True, null=True)
